+++ jetpack_vendor/i18n-map.php
@@ -4,9 +4,21 @@ return array(
   'domain' => 'zero-bs-crm',
   'type' => 'plugins',
   'packages' => array(
+    'jetpack-admin-ui' => array(
+      'path' => 'jetpack_vendor/automattic/jetpack-admin-ui',
+      'ver' => '0.2.20',
+    ),
     'jetpack-assets' => array(
       'path' => 'jetpack_vendor/automattic/jetpack-assets',
       'ver' => '1.18.3',
+    ),
+    'jetpack-connection' => array(
+      'path' => 'jetpack_vendor/automattic/jetpack-connection',
+      'ver' => '1.51.10-alpha',
+    ),
+    'jetpack-forms' => array(
+      'path' => 'jetpack_vendor/automattic/jetpack-forms',
+      'ver' => '0.17.1-alpha',
     ),
   ),
 );